requirements.txt
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
python-dotenv==1.0.0
google-generativeai==0.8.5
azure-ai-formrecognizer==3.3.0
Werkzeug==3.1.3
requests==2.32.4
pillow==11.3.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.8.0
bcrypt==4.2.0
pdfplumber 
langchain 
langchain-google-genai 
langchain-community 
faiss-cpu 
google-generativeai
